Async version of the nextcloud EventBus.
The idea is to be able to use

await emit(stuff...);

in order to wait for completion of the event handlers. This is a
feature. Just do not await the result and it should work just like
before.

We also define typings for the expected return types.

New unsubscribe() and unsubscribeAll() methods.